fix #275
