+++ example/lib/screens/common/example_button.dart
@@ -1,0 +1,43 @@
+import 'package:flutter/material.dart';
+
+class ExampleButtonNode extends StatelessWidget {
+  const ExampleButtonNode({
+    this.title,
+    this.onPressed,
+  });
+
+  final String title;
+  final Function onPressed;
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      margin: const EdgeInsets.symmetric(
+        vertical: 20.0,
+      ),
+      child: Column(
+        children: <Widget>[
+          Text(
+            title,
+            textAlign: TextAlign.center,
+            style: const TextStyle(
+              color: Colors.black,
+              fontSize: 21.0,
+              fontWeight: FontWeight.w600,
+            ),
+          ),
+          Container(
+            margin: const EdgeInsets.only(
+              top: 10.0,
+            ),
+            child: RaisedButton(
+              onPressed: onPressed,
+              child: const Text("Open example"),
+              color: Colors.amber,
+            ),
+          )
+        ],
+      ),
+    );
+  }
+}
